Add docker image which includes pre-stop hook script.
This adds a container image derived from the Debian coturn image which includes
a pre-stop hook script. This hook script polls the Prometheus metrics endpoint
and waits until the number of active allocations reaches zero. This can be used
to detect when traffic is drained from a running coturn instance, in order to
perform a non-invasive graceful restart.